test(button + link): add helper methods to create AllVariants stories
