--- packages/components/src/Button/button.stories.module.css
@@ -1,15 +1,0 @@
-.color {
-  @apply p-2;
-}
-
-.variant {
-  @apply mb-5;
-}
-
-.iconLink {
-  @apply ml-1;
-}
-
-.iconButton {
-  @apply ml-2;
-}
