+++ packages/components/src/Button/button.stories.tsx
@@ -1,21 +1,13 @@
 import { Story } from "@storybook/react/types-6-0";
 import React from "react";
-import { getRecommendedVariants } from "../../.storybook/buttonUtils";
-import ClickableStyle from "../ClickableStyle";
-import Heading from "../Heading";
-import AddRoundedIcon from "../Icons/AddRounded";
-import Text from "../Text";
-import Button, { ButtonProps } from "./button";
-import styles from "./button.stories.module.css";
+import {
+  getRecommendedVariants,
+  getAllVariantsWithStates,
+  getLargeVariantsOnDarkBackgroundWithStates,
+} from "../../.storybook/buttonUtils";
+import Button from "./button";
 
-const sizes = ["large", "medium", "small"] as const;
-const allColors = ["alert", "brand", "neutral", "success", "warning"] as const;
-// "link" is ommitted here because it's rendered separately since it only has one size
-const variants = ["flat", "outline", "plain"] as const;
-const states = ["inactive", "hover", "focus", "active", "disabled"] as const;
-
-// For now, the UI kit only includes alert & brand "flat" buttons
-const flatColors = ["alert", "brand"] as const;
+const colors = ["alert", "brand", "neutral", "success", "warning"] as const;
 
 export default {
   title: "Button",
@@ -29,7 +21,7 @@ export default {
     color: {
       control: {
         type: "radio",
-        options: allColors,
+        options: colors,
       },
     },
   },
@@ -47,92 +39,23 @@ Primary.args = {
 export const RecommendedVariants = () =>
   getRecommendedVariants(Button, "Button");
 
-// Show grids with all variants
-
 const gridParameters = {
   axe: {
     skip: true,
   },
+  // Skip snapshots because this grid has too many components with all the states,
+  // and the previous stories generate enough for our needs.
+  // These stories are more for visual regression testing.
   snapshot: {
     skip: true,
   },
 };
 
-const renderVariant = (
-  variant: ButtonProps["variant"],
-  headingColor: "white" | "neutral",
-  buttonChildren: React.ReactNode,
-  size?: ButtonProps["size"],
-) => {
-  const colors = variant === "flat" ? flatColors : allColors;
-  const icon = variant === "plain" && (
-    <AddRoundedIcon className={styles.iconButton} purpose="decorative" />
-  );
-
-  return (
-    <React.Fragment key={variant}>
-      <Heading size="h2" color={headingColor}>
-        {variant}
-        {size && ` - ${size}`}
-      </Heading>
-      <table className={styles.variant}>
-        <tbody>
-          {states.map((state) => (
-            <tr key={state}>
-              <th scope="row">
-                <Text size="body" color={headingColor}>
-                  {state}
-                </Text>
-              </th>
-              {colors.map((color) => (
-                <td key={color} className={styles.color}>
-                  {/* To pass the "state" prop (only used for demonstration in storybook),
-                  we must use ClickableStyle instead of Button */}
-                  <ClickableStyle
-                    as={"button"}
-                    size={size}
-                    color={color}
-                    variant={variant}
-                    state={state}
-                    disabled={state === "disabled"}
-                  >
-                    {buttonChildren}
-                    {icon}
-                  </ClickableStyle>
-                </td>
-              ))}
-            </tr>
-          ))}
-        </tbody>
-      </table>
-    </React.Fragment>
-  );
-};
-
-export const AllVariants = () => (
-  <ul>
-    <li>{renderVariant("link", "neutral", "Button")}</li>
-    {sizes.map((size) => (
-      <li key={size}>
-        {variants.map((variant) =>
-          renderVariant(variant, "neutral", "Button", size),
-        )}
-      </li>
-    ))}
-  </ul>
-);
+export const AllVariants = () => getAllVariantsWithStates("button", "Button");
 AllVariants.parameters = gridParameters;
 
-export const LargeVariantsOnDarkBackground = () => (
-  <ul>
-    <li key="link">{renderVariant("link", "white", "Button")}</li>
-    {variants.map((variant) => (
-      <li key={variant}>
-        {renderVariant(variant, "white", "Button", "large")}
-      </li>
-    ))}
-  </ul>
-);
+export const LargeVariantsOnDarkBackground = () =>
+  getLargeVariantsOnDarkBackgroundWithStates("button", "Button");
 LargeVariantsOnDarkBackground.parameters = {
   ...gridParameters,
   backgrounds: {
