+++ packages/components/src/Link/Link.stories.tsx
@@ -1,22 +1,14 @@
 import { Story } from "@storybook/react/types-6-0";
 import React from "react";
-import { getRecommendedVariants } from "../../.storybook/buttonUtils";
-import ClickableStyle from "../ClickableStyle";
-import Heading from "../Heading";
-import AddRoundedIcon from "../Icons/AddRounded";
+import {
+  getRecommendedVariants,
+  getAllVariantsWithStates,
+  getLargeVariantsOnDarkBackgroundWithStates,
+} from "../../.storybook/buttonUtils";
 import Text from "../Text";
-import Link, { LinkProps } from "./Link";
-import styles from "./Link.stories.module.css";
+import Link from "./Link";
 
-const sizes = ["large", "medium", "small"] as const;
-const allColors = ["alert", "brand", "neutral", "success", "warning"] as const;
-// "link" is ommitted here because it's rendered separately since it only has one size
-const variants = ["flat", "outline", "plain"] as const;
-// "disabled" is ommitted here because <a> tags can't actually be disabled
-const states = ["inactive", "hover", "focus", "active"] as const;
-
-// For now, the UI kit only includes alert & brand "flat" buttons.
-const flatColors = ["alert", "brand"] as const;
+const colors = ["alert", "brand", "neutral", "success", "warning"] as const;
 
 export default {
   title: "Link",
@@ -30,7 +22,7 @@ export default {
     color: {
       control: {
         type: "radio",
-        options: allColors,
+        options: colors,
       },
     },
   },
@@ -74,91 +66,24 @@ linkInHeading.args = {
   ...defaultArgs,
 };
 
-// Show grids with all variants
-
 const gridParameters = {
   axe: {
     skip: true,
   },
+  // Skip snapshots because this grid has too many components with all the states,
+  // and the previous stories generate enough for our needs.
+  // These stories are more for visual regression testing.
   snapshot: {
     skip: true,
   },
 };
 
-const renderVariant = (
-  variant: LinkProps["variant"],
-  headingColor: "white" | "neutral",
-  buttonChildren: React.ReactNode,
-  size?: LinkProps["size"],
-) => {
-  const colors = variant === "flat" ? flatColors : allColors;
-  const icon = variant === "plain" && (
-    <AddRoundedIcon className={styles.iconButton} purpose="decorative" />
-  );
+export const AllVariants = () => getAllVariantsWithStates("a", "Link");
+AllVariants.parameters = gridParameters;
 
-  return (
-    <React.Fragment key={variant}>
-      <Heading size="h2" color={headingColor}>
-        {variant}
-        {size && ` - ${size}`}
-      </Heading>
-      <table className={styles.variant}>
-        <tbody>
-          {states.map((state) => (
-            <tr key={state}>
-              <th scope="row">
-                <Text size="body" color={headingColor}>
-                  {state}
-                </Text>
-              </th>
-              {colors.map((color) => (
-                <td key={color} className={styles.color}>
-                  {/* To pass the "state" prop (only used for demonstration in storybook),
-                  we must use ClickableStyle instead of Link */}
-                  <ClickableStyle
-                    as={"a"}
-                    size={size}
-                    color={color}
-                    variant={variant}
-                    state={state}
-                    href=""
-                  >
-                    {buttonChildren}
-                    {icon}
-                  </ClickableStyle>
-                </td>
-              ))}
-            </tr>
-          ))}
-        </tbody>
-      </table>
-    </React.Fragment>
-  );
-};
-
-export const allVariants = () => (
-  <ul>
-    <li>{renderVariant("link", "neutral", "Link")}</li>
-    {sizes.map((size) => (
-      <li key={size}>
-        {variants.map((variant) =>
-          renderVariant(variant, "neutral", "Link", size),
-        )}
-      </li>
-    ))}
-  </ul>
-);
-allVariants.parameters = gridParameters;
-
-export const largeVariantsOnDarkBackground = () => (
-  <ul>
-    <li key="link">{renderVariant("link", "white", "Link")}</li>
-    {variants.map((variant) => (
-      <li key={variant}>{renderVariant(variant, "white", "Link", "large")}</li>
-    ))}
-  </ul>
-);
-largeVariantsOnDarkBackground.parameters = {
+export const LargeVariantsOnDarkBackground = () =>
+  getLargeVariantsOnDarkBackgroundWithStates("a", "Link");
+LargeVariantsOnDarkBackground.parameters = {
   ...gridParameters,
   backgrounds: {
     default: "dark",
